Instantiate the worker's intermediate results queue in __init__
remove unused file
